requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
fastapi>=0.109.0
uvicorn>=0.27.0
google-api-python-client>=2.114.0
google-auth>=2.27.0
anthropic>=0.18.0
pytest>=8.0.0
